[fetch-later] Define new permissions policy `deferred-fetch`
1-Pager: https://docs.google.com/document/d/1P70kdENIByy3qWabN5rUPmBVkkANNSOM_jZynLnqINY/edit

This CL follows the [guide] to define a new permissions policy
`deferred-fetch`, which is used to gate the new `fetchLater()` API.
Relevant WPT are added in this CL, and the subsequent CL will further
use this policy to adjust the request quota.

In this CL, `deferred-fetch` is not added to chrome://settings/content
page. And no permission prompt for it.

- webappsec request: https://github.com/w3c/webappsec-permissions-policy/issues/544
- "deferred-fetch" usage in fetchLater draft spec: https://whatpr.org/fetch/1647.html#request-a-deferred-fetch
- "deferred-fetch" policy discussion: https://github.com/WICG/pending-beacon/issues/87#issuecomment-1985358609
- FetchLater Explainer: https://github.com/WICG/pending-beacon/blob/main/docs/fetch-later-api.md

[guide]: https://chromium.googlesource.com/chromium/src/+/main/components/permissions/add_new_permission.md

Bug: b:40276121
Change-Id: I2db4dd7484610fefb50c463552155b220f13ac5d