[layout] Fix synthesis of inline baselines.
Previously for <button>, <input type=button> (and friends), and
<input type=range> we'd explicitly produce a baseline at appropriate
"end" edge.

This wasn't strictly correct instead we should be synthesizing them
within LogicalBoxFragment::BaselineMetrics.

There are three different ways to synthesize these baseline, off the
margin-box (the default), content-box (buttons), and border-box (range
and checkbox/radio with effective appearance).

Instead of checking all these conditions, this patch introduces an
additional field to the computed style to tell BaselineMetrics which
edge the synthesize off.

This patch makes changes two WPT tests which (I believe) to be
incorrect see:
https://github.com/web-platform-tests/wpt/issues/45889

The TL;DR is that:
<button></button> <button style="contain:layout"></button>

Currently have different baselines when they shouldn't.

The image rebaselines are ensuring that we synthesize the baseline
for range controls consistently. Previously we'd synthesize them off
the border-box if they were in the same writing-mode, and off the
margin-box if in a different writing-mode.

Bug: 40937312
Change-Id: I28eedfbce58ead9d5e3119ae3b164c559fb9267c